fix edit grant
