+++ deploy-web/src/pages/settings/authorizations.tsx
@@ -3,22 +3,19 @@ import { NextSeo } from "next-seo";
 import PageContainer from "@src/components/shared/PageContainer";
 import SettingsLayout, { SettingsTabs } from "@src/components/settings/SettingsLayout";
 import { Fieldset } from "@src/components/shared/Fieldset";
-import { Box, Button, CircularProgress, IconButton, Table, TableBody, TableCell, TableContainer, TableRow } from "@mui/material";
+import { Box, Button, CircularProgress, Table, TableBody, TableCell, TableContainer, TableRow } from "@mui/material";
 import { useState } from "react";
 import { useKeplr } from "@src/context/KeplrWalletProvider";
-import { CustomTableHeader, CustomTableRow } from "@src/components/shared/CustomTable";
-import { AKTAmount } from "@src/components/shared/AKTAmount";
-import { coinToUDenom } from "@src/utils/priceUtils";
-import { FormattedTime } from "react-intl";
+import { CustomTableHeader } from "@src/components/shared/CustomTable";
 import { Address } from "@src/components/shared/Address";
 import { GrantModal } from "@src/components/wallet/GrantModal";
-import EditIcon from "@mui/icons-material/Edit";
-import DeleteIcon from "@mui/icons-material/Delete";
 import { GrantType } from "@src/types/grant";
 import { TransactionMessageData } from "@src/utils/TransactionMessageData";
 import AccountBalanceIcon from "@mui/icons-material/AccountBalance";
 import { useGranteeGrants, useGranterGrants } from "@src/queries/useGrantsQuery";
 import { Popup } from "@src/components/shared/Popup";
+import { GranterRow } from "@src/components/settings/GranterRow";
+import { GranteeRow } from "@src/components/settings/GranteeRow";
 
 type Props = {};
 
@@ -95,25 +92,7 @@ const SettingsSecurityPage: React.FunctionComponent<Props> = ({}) => {
 
                       <TableBody>
                         {granterGrants.map(grant => (
-                          <CustomTableRow key={grant.grantee}>
-                            <TableCell>
-                              <Address address={grant.grantee} isCopyable />
-                            </TableCell>
-                            <TableCell align="right">
-                              <AKTAmount uakt={coinToUDenom(grant.authorization.spend_limit)} /> AKT
-                            </TableCell>
-                            <TableCell align="right">
-                              <FormattedTime year="numeric" month={"numeric"} day={"numeric"} value={grant.expiration} />
-                            </TableCell>
-                            <TableCell>
-                              <IconButton onClick={() => onEditGrant(grant)}>
-                                <EditIcon />
-                              </IconButton>
-                              <IconButton onClick={() => setDeletingGrant(grant)}>
-                                <DeleteIcon />
-                              </IconButton>
-                            </TableCell>
-                          </CustomTableRow>
+                          <GranterRow key={grant.grantee} grant={grant} onEditGrant={onEditGrant} setDeletingGrant={setDeletingGrant} />
                         ))}
                       </TableBody>
                     </Table>
@@ -143,17 +122,7 @@ const SettingsSecurityPage: React.FunctionComponent<Props> = ({}) => {
 
                       <TableBody>
                         {granteeGrants.map(grant => (
-                          <CustomTableRow key={grant.granter}>
-                            <TableCell>
-                              <Address address={grant.granter} isCopyable />
-                            </TableCell>
-                            <TableCell align="right">
-                              <AKTAmount uakt={coinToUDenom(grant.authorization.spend_limit)} /> AKT
-                            </TableCell>
-                            <TableCell align="right">
-                              <FormattedTime year="numeric" month={"numeric"} day={"numeric"} value={grant.expiration} />
-                            </TableCell>
-                          </CustomTableRow>
+                          <GranteeRow key={grant.granter} grant={grant} />
                         ))}
                       </TableBody>
                     </Table>
@@ -173,6 +142,7 @@ const SettingsSecurityPage: React.FunctionComponent<Props> = ({}) => {
               onClose={() => setDeletingGrant(null)}
               onCancel={() => setDeletingGrant(null)}
               onValidate={onDeleteGrantConfirmed}
+              enableCloseOnBackdropClick
             >
               Deleting grant to{" "}
               <strong>
