+++ deploy-web/src/types/network.ts
@@ -1,0 +1,12 @@
+export type Network = {
+  id: string;
+  title: string;
+  description: string;
+  nodesUrl: string;
+  chainId: string;
+  versionUrl: string;
+  rpcEndpoint?: string;
+  version: string;
+  enabled: boolean;
+  suggestKeplrChain?: () => Promise<void>;
+};
